backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.0.0
python-dotenv>=1.0.0
opencv-python-headless
numpy<2.0.0
python-multipart
Pillow>=10.0.0
SQLAlchemy>=2.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
pyzbar>=0.1.9
